Rename RELEASE-NOTES.txt to RELEASE-NOTES.md
In #1576 we renamed `CHANGELOG.md` to `RELEASE-NOTES.txt` and in so doing we broke any integration which attempted to render the contents of the file as Markdown. The file continued to be written as a Markdown document but by renaming the extension we prevented any tools which rely on the extension to detect the type and so valuable tools like GitHub's diff viewer wouldn't show the contents of the file as they were written but instead as if they were plain text.

In this patch we're renaming it back to a `.md` extension to reflect the reality of the file that it is indeed Markdown content. In so doing we're fixing the integrations that were broken in #1576.