Update `node-simperium` to fix some duplicate-write issues
In certain situations we can find more than one WebSocket connection to
Simperium open and messages come back from the server over both. This
tends to happen when the network connection is spotty: `node-simperium`
thinks the WebSocket is closed but the browser still flushes out the
last messages on it as soon as network connectivity returns.

In the update to `node-simperium` it guards against certain kinds of
"ghost writing" in these situations by making sure that duplicate or
out-of-order messages from the server don't get applied more than once.