Disable note-editing from list view and open note on Enter presses
It was possible to edit the contents of a note in a small-screen view
with the note list open. This was confusing because it wasn't obvious
that the edits were applying.

In this change the editor (with help of some ugly prop-drilling)
disables itself while in a small-screen view while the note list is open.
In order to ease the jump back to the note the Enter key now opens it.