Hotkeys: Always allow using the Ctrl-T shortcut to jump between note/tags
Previously we were only allowing a switch _from_ the note editor _to_ the
tag editor, or _from_ the tag editor _to_ the note editor, but we really
want to allow jumping between these fields any time, even if none is
selected.

With the work to improve keyboard shortcut usability this is more important
since we're emphasizing keyboard navigation.

In this patch we change the conditional from one that gates the behavior
to one that uses the current focus to determine which to jump to.