Stop rendering editor when it's not visible
Previously we were always rendering the editor.
When very large or slow notes were in the list it made for
abrupt pauses while scrolling through the list.

In this change we stop rendering the editor altogether if
only the list view is visible. This has removed all abrupt
pauses when scrolling through the list view on the
narrow-screen mode.