Fix note list toggle (#2037)
* most of a fix for this by not overloading CLOSE_NOTE

* default to true so we show the note list on first load on small screens

* toggle back into the note list after next/prev

* don't open note list on next/prev if it's not already open

* change wording of note list key since it's a toggle now

* Remove some unused code

* Remove toggling behavior from note list

* Separate the action of selecting a note from the action of opening it

* Disable note-editing from list view and open note on Enter presses

It was possible to edit the contents of a note in a small-screen view
with the note list open. This was confusing because it wasn't obvious
that the edits were applying.

In this change the editor (with help of some ugly prop-drilling)
disables itself while in a small-screen view while the note list is open.
In order to ease the jump back to the note the Enter key now opens it.

* review cleanup

* Keep highlighted note in view in list view

* Stop rendering editor when it's not visible

Previously we were always rendering the editor.
When very large or slow notes were in the list it made for
abrupt pauses while scrolling through the list.

In this change we stop rendering the editor altogether if
only the list view is visible. This has removed all abrupt
pauses when scrolling through the list view on the
narrow-screen mode.

* Remove workaround for disabling the editor while hidden

Co-authored-by: Dennis Snell <dennis.snell@automattic.com>