[Rewrite Branch] Fix printing of notes (#2271)
This PR adds a display none div that contains the content of a note. When the user prints it displays that div and hides other content on the page.

Adding a display none div with the content allows us to do this purely with CSS rather than watching for print events and potentially running into race conditions with React rendering.