Fix preview note view: grab proper note object and update with props
In #2281 we fixed the note preview so that it would only render a
Markdown-enabled note if the preview mode in the editor is set.

Unfortunately we missed a dependency in the `useEffect()` callback
and so the preview would only render the corresponding view to the
settings passed when it was first opened. This has been fixed by
adding the proper dependency.

Additionally we were only getting the Markdown setting for the passed
note, such as when viewing revisions from a note's history. For normal
preview views we missed examining the opened note. We have fixed this
by picking the appropriate note object to inspect when determining
which mode to render.