Adds highlighting in note for search query (#2292)
* Add document highlight provider to highglight search results

In this patch we're adding a new document highlight provider to highlight the
search results in a note. This may be a useful approach because it relies
entirely on Monaco's API to provide the functionality. Unfortunately in my
exploration it wasn't clear to me to how get this to highlight when the cursor
isn't on any word. The API appears to be designed to highlight related words to
the word under the cursor. If the cursor is on an empty line then nothing else
gets highlighted and the provider never gets called.

If we can force-trigger the highlighting to run this might be workable. We need
to do that and also be able to select an active highlight if we want to use
this for search.

We might have `editor.action.wordHighlight.trigger` as an action to call but I
gave up trying to figure out how to properly call it.

* Change highlight colors

* Use decorations

* Fix css class

* Remove uneeded highlighting styles

* Combine title and search decorators

* Make occurrencesHighlight false again

* Account for notes with no content

* Make highlight text white when in light mode

* Cleanup from code review

* Gate this to monaco so it doesnt run on fast view

Co-authored-by: Dennis Snell <dennis.snell@automattic.com>