Fixes the css so all pages print (#2319)
There were three issues with printing. The position relative shows more than the first page. The overflow auto allows all the pages to print. I have a note that should bee 22 pages but only 11 were printed without the overflow auto. The last issue is that display flex in Firefox causes only the first page to print. So added display: block to fix that.