Change tag autocomplete logic to make it work (#2350)
This updates the logic to get the correct caret (cursor) position inside the contentEditable. The previous logic did not provide the current cursor position.

endOffset is not what you would expect, it does not give you the cursor position. The selection starts and ends at the same place. You have to get the contents of the element and determine where you are within that element.

This PR additionally returns early if window.getSelection does not exist.