Add Electron close dialog (#2277)
When we close the app we want to ensure that the user knows if there are unsynchronized notes. Closing the app will cause those changes to sit unsynchronized and could lead to data loss.

This PR utilizes the work done on the Log out unsynchronized notes dialog. When the user clicks "Quit Simplenote" in the system menu or uses the close button on the window we send an event to the app to check if there are unsynchronized notes. If there are no unsynchronized notes we send an event back to electron to close the application. If there are unsynchronized notes we pop the confirmation dialog. If the click the continue button we will send an event to electron to close the app.