Display cannonical lexical tag name (#2435)
When displaying tags on a note this PR uses the canonical lexical version on the Tag entity and not the lexical tag name on the note. Additionally, this PR makes it so during a rename we no longer change the tags on a note if it is just a lexical change.