When there are no notes suggest creating one (#2422)
When a user does not have any notes suggest adding a new note. Additionally, if a user types a search term that does not match any notes suggest creating a new not with that query.

This also fixes a bug where if you have no notes "Notes Loading" neve disappears.