Create welcome note via Redux to ensure search indexing
Previously the welcome note would not appear in the app until a refresh.
This was caused because the search index didn't see it, as it was being
created directly in Simperium without a corresponding Redux action.

In this patch we're creating the welcome note with a dispatch via Redux
which ensures that it's immeidately indexed by the search. We dispatch
inside the middleware setup which should cause a problem, except that
we have a Promise on importing the content of the weleome note and this
`then()` ensures the code only runs after the middleware has loaded.

It could happen that if the search middleware weren't yet loaded then
we could miss the welcome note again, but since we're using `import()`
which loads the module (presumably from a network call) then we should
have no practical worry about timing issues.