Update tag submission from note editor (#2607)
When adding tags from the note editor you can add spaces and then when you press enter or a comma tags are submitted and separated on the spaces. This is brought up in #2207.

In addition when entering a tag if you switch to another note, or click out of the tag input area the unfinished tag stays as described in #2589

This PR will submit the new tag when you enter a space or leave the tag input area.