Updating the ui for the editing tags section (#2584)
As part of the UI Spring cleaning we are doing this looks to take care of updating the Edit Tags section. Resolves #2567.

This changes from using a cross icon on the left side of the tag name to using a trash icon on the right side.
While editing the styles I also removed a number of lines that don't appear to be in use anymore at all.

I also updated some of the styles to better use the classes in the themes.scss file instead of colors in the component styles. Not all are removed more refactoring will be needed.

Removed two icons no longer needed cross-outline-small and cross-outline.