Fix scroll bar styles in Firefox (#2641)
This PR is to resolve #2640.
In Firefox we were using the default styling for scroll bars in areas like the note list. In some cases, this caused a scroll bar that didn't match the rest of the app.
Particularly on MacOS with System Preferences > General > Show scroll bars: set to anything other than When scrolling and using dark mode.