Update the navigation bar and taglist styles (#2649)
As part of the UI spring cleanup project, this PR works to resolve #2566.

This causes the navigation bar to open on top of other content rather than push it to the side.
It updates the styles and colors used.
It also adds an overlay to the other content when the navigation bar is open instead of reducing opacity.