Make keyboard shortcuts for searching always active regardless of keyboard shortcut preference (#2652)
This PR resolves #2651.
Currently, the browser's built-in search doesn't work well with the note editor. For content not in view in the editor, the search is not able to find and select matching content.
For this reason, we are going to keep the keyboard shortcut for searching always on regardless of the keyboard shortcut preference.