Fixing duplicate file name issue in exports (#2694)
When trying to export notes there are some instances where notes with the same title will overwrite each other. This is caused by some characters in the titles that can break our method for checking for duplicate file names.
The root cause identified (lodash.update was interpreting the "filename" as a property path, thus interpreting. and array-like indexing with brackets. This removes the update function and uses a Map to check for duplicates instead.