Parse multiple, pasted tags
In order to avoid a lot of weird edge cases that arise when pasting
multiple tags, this now converts any portion of the string that is
separated by a "submit" character into a submitted tag. The trailing
value is retained as a pending tag that a user may edit before
submitting.