Parse multiple pasted tags and edit pending tags (#2756)
* Allow modifying pending tags

Previously, the caret was aggressively moved to the end of the tag input
regardless of whether the user attempted to click elsewhere or insert
text at a different location. These changes retain a reference to the
last known cursor position and restore it during relevant interactions.

* Parse multiple, pasted tags

In order to avoid a lot of weird edge cases that arise when pasting
multiple tags, this now converts any portion of the string that is
separated by a "submit" character into a submitted tag. The trailing
value is retained as a pending tag that a user may edit before
submitting.