Set a body data attribute for the theme selection (#2873)
This works to update the way we set the theme being used in the app.
The current method has been to set a class such as theme-light or theme-dark in a number of places, the main one a couple of elements below the #root div of the app.

This had the added effect of not reaching modals opened as they were as part of the body element. This meant each modal had to have the class added as well.

This approach adds a data attribute to the body element such as data-theme="dark" or data-theme="light" instead.

There should be no user-facing changes here.