Add system theme for Windows and Mac Electron builds (#2882)
Picked up from PR #2135

Currently in our Electron builds we don't offer a system theme that responds to the OSs color preference but is available in the web app.

This enables it for both Windows and Mac builds. There appears to be a bug in Electron or Chromium on Linux that the app doesn't respond properly to changing the color preferences so it still won't work there. Possibly this will get resolved in Electron 13 due to come out May 25, 2021.

It also updates so that the system menu for Windows updates to be the choice of theme in Simplenote itself. So if you specifically choose Light the menu will be light, same for dark, overriding the system color preference.

This semi addresses issues #2134 and #1799 as they will both be resolved for Mac and Windows, but still not working on the Linux builds.