Make search global (#2878)
Fixes #2626
This updates the way search is handled if there is an opened tag to match the other platforms.
Currently, if there is a tag selected the search is restricted to that tag.
This updates so that whenever a search is placed it is for all notes. Searches can still be restricted to tags using the syntax tag:recipe in the search field.