Fixes a bug where you had to use CMD+Q twice to quit the mac app. (#2901)
Fixes #2676

In Mac OS there the expected behavior is that when you close the window of an app that it stays running and you can open it again from the dock. Then if you quit the app using ⌘Q it completely closes.

Currently in order to fully quit you need to press ⌘Q twice which is unexpected. This resolves that by allowing you to close the window and reopen it, but also quit the app as expected.