Update display of logo for windows start menu tile (#2937)
Fixes #2935

Currently, if you install the appx or Windows store version of Simplenote and then pin it to the start menu the tile generated has a large logo and doesn't show the app name at all. This is different from most other apps installed.

This shrinks the size of the logo and displays the name on the tile.