Update Dark mode colors to use to css variables (#2936)
This is a continuation of previous work to update the CSS to use variables for colors.
This moves all colors for dark mode to use CSS variables instead.

There should be no user impact or any noticeable changes. It's an intermediate step to simplifying the CSS used throughout the app.