Add handling of 429 HTTP errors on authentication (#3029)
There are times when rate limiting will be applied when trying to login multiple times unsuccessfully causing future login attempts to fail as well. Currently, there is no way to show this in the app leaving things confusing if you get rate limited. This updates to handle that error case when logging in.