Temporarily remove `appx` target—Will `exe` sign okay?
The `appx` was built before the `exe`, but with the latest config
changes it failed to code sign.

https://buildkite.com/automattic/simplenote-electron/builds/173#0190e360-f70d-4ee9-8fd7-15d8cff032d4

At the same time, with the config before those changes, the `exe` was
not signed, or not signed properly. Regardless, Windows did not want to
open that file.

Removing `appx` to get to the `exe` build in CI and see if it signs okay
with these settings